Overwrite latest result in history graph
After 20 checks, the graph wasn't obviously "pushed" to the left.

Now, having the end of the graph "wiped" with a space, it can be seen that the graph was shift to the left.

Example (recorded after the 3rd host was checked):
See that the bottom 3 hosts weren't checked so far (a space is seen at the end of the graph).

Host address   Current [ms]   History graph            Success   Avg [ms]  Rcvd#
127.0.0.1         0.3910      .....................    100.0%       0.3844    42
127.0.0.2         0.2689      .....................    100.0%       0.3194    42
127.0.0.3         0.2849      .....................    100.0%       0.3252    42
8.8.8.8          10.2119      ....................     100.0%      10.2455    41
127.0.0.4         0.3650      ....................     100.0%       0.3342    41
172.16.0.1                    --------------------       0.0%                  0